+++ tests/git/test_git_commands.py
@@ -246,9 +246,19 @@ def test_setup_no_secret_existing_api(
     )
 
 
+@pytest.mark.parametrize(
+    "repo_name, int_name, secret_name",
+    [
+        ("db.schema.FooRepo", "FooRepo_api_integration", "db.schema.FooRepo_secret"),
+        ("schema.FooRepo", "FooRepo_api_integration", "schema.FooRepo_secret"),
+        ("FooRepo", "FooRepo_api_integration", "FooRepo_secret"),
+    ],
+)
 @mock.patch("snowflake.connector.connect")
 @mock.patch("snowflake.cli._plugins.snowpark.commands.ObjectManager.describe")
-def test_setup_no_secret_create_api(mock_om_describe, mock_connector, runner, mock_ctx):
+def test_setup_no_secret_create_api(
+    mock_om_describe, mock_connector, runner, mock_ctx, repo_name, int_name, secret_name
+):
     mock_om_describe.side_effect = ProgrammingError(
         errno=DOES_NOT_EXIST_OR_NOT_AUTHORIZED
     )
@@ -256,7 +266,7 @@ def test_setup_no_secret_create_api(mock_om_describe, mock_connector, runner, mo
     mock_connector.return_value = ctx
 
     communication = "\n".join([EXAMPLE_URL, "n", "", ""])
-    result = runner.invoke(["git", "setup", "repo_name"], input=communication)
+    result = runner.invoke(["git", "setup", repo_name], input=communication)
 
     assert result.exit_code == 0, result.output
     assert result.output.startswith(
@@ -264,22 +274,22 @@ def test_setup_no_secret_create_api(mock_om_describe, mock_connector, runner, mo
             [
                 "Origin url: https://github.com/an-example-repo.git",
                 "Use secret for authentication? [y/N]: n",
-                "API integration identifier (will be created if not exists) [repo_name_api_integration]: ",
-                "API integration 'repo_name_api_integration' successfully created.",
+                f"API integration identifier (will be created if not exists) [{int_name}]: ",
+                f"API integration '{int_name}' successfully created.",
             ]
         )
     )
     assert ctx.get_query() == dedent(
-        """
-        create api integration IDENTIFIER('repo_name_api_integration')
+        f"""
+        create api integration IDENTIFIER('{int_name}')
         api_provider = git_https_api
         api_allowed_prefixes = ('https://github.com/an-example-repo.git')
         allowed_authentication_secrets = ()
         enabled = true
         
         
-        create git repository IDENTIFIER('repo_name')
-        api_integration = repo_name_api_integration
+        create git repository IDENTIFIER('{repo_name}')
+        api_integration = {int_name}
         origin = 'https://github.com/an-example-repo.git'
         """
     )
@@ -327,10 +337,18 @@ def test_setup_existing_secret_existing_api(
     )
 
 
+@pytest.mark.parametrize(
+    "repo_name, int_name, secret_name",
+    [
+        ("db.schema.FooRepo", "FooRepo_api_integration", "db.schema.FooRepo_secret"),
+        ("schema.FooRepo", "FooRepo_api_integration", "schema.FooRepo_secret"),
+        ("FooRepo", "FooRepo_api_integration", "FooRepo_secret"),
+    ],
+)
 @mock.patch("snowflake.connector.connect")
 @mock.patch("snowflake.cli._plugins.snowpark.commands.ObjectManager.describe")
 def test_setup_existing_secret_create_api(
-    mock_om_describe, mock_connector, runner, mock_ctx
+    mock_om_describe, mock_connector, runner, mock_ctx, repo_name, int_name, secret_name
 ):
     mock_om_describe.side_effect = [
         ProgrammingError(errno=DOES_NOT_EXIST_OR_NOT_AUTHORIZED),
@@ -342,7 +360,7 @@ def test_setup_existing_secret_create_api(
     mock_connector.return_value = ctx
 
     communication = "\n".join([EXAMPLE_URL, "y", "existing_secret", "", ""])
-    result = runner.invoke(["git", "setup", "repo_name"], input=communication)
+    result = runner.invoke(["git", "setup", repo_name], input=communication)
 
     assert result.exit_code == 0, result.output
     assert result.output.startswith(
@@ -350,24 +368,24 @@ def test_setup_existing_secret_create_api(
             [
                 "Origin url: https://github.com/an-example-repo.git",
                 "Use secret for authentication? [y/N]: y",
-                "Secret identifier (will be created if not exists) [repo_name_secret]: existing_secret",
+                f"Secret identifier (will be created if not exists) [{secret_name}]: existing_secret",
                 "Using existing secret 'existing_secret'",
-                "API integration identifier (will be created if not exists) [repo_name_api_integration]: ",
-                "API integration 'repo_name_api_integration' successfully created.",
+                f"API integration identifier (will be created if not exists) [{int_name}]: ",
+                f"API integration '{int_name}' successfully created.",
             ]
         )
     )
     assert ctx.get_query() == dedent(
-        """
-        create api integration IDENTIFIER('repo_name_api_integration')
+        f"""
+        create api integration IDENTIFIER('{int_name}')
         api_provider = git_https_api
         api_allowed_prefixes = ('https://github.com/an-example-repo.git')
         allowed_authentication_secrets = (existing_secret)
         enabled = true
 
 
-        create git repository IDENTIFIER('repo_name')
-        api_integration = repo_name_api_integration
+        create git repository IDENTIFIER('{repo_name}')
+        api_integration = {int_name}
         origin = 'https://github.com/an-example-repo.git'
         git_credentials = existing_secret
         """
@@ -435,25 +453,35 @@ def test_setup_create_secret_create_api(
         (
             "@repo/branches/main/",
             "@repo/branches/main/",
-            ["@repo/branches/main/s1.sql", "@repo/branches/main/a/s3.sql"],
+            ["@repo/branches/main/s1.sql", "@repo/branches/main/a/S3.sql"],
         ),
         (
             "@repo/branches/main/a",
             "@repo/branches/main/",
-            ["@repo/branches/main/a/s3.sql"],
+            ["@repo/branches/main/a/S3.sql"],
         ),
         (
             "@db.schema.repo/branches/main/",
             "@db.schema.repo/branches/main/",
             [
                 "@db.schema.repo/branches/main/s1.sql",
-                "@db.schema.repo/branches/main/a/s3.sql",
+                "@db.schema.repo/branches/main/a/S3.sql",
             ],
         ),
         (
             "@db.schema.repo/branches/main/s1.sql",
             "@db.schema.repo/branches/main/",
             ["@db.schema.repo/branches/main/s1.sql"],
+        ),
+        (
+            "@DB.SCHEMA.REPO/branches/main/s1.sql",
+            "@DB.SCHEMA.REPO/branches/main/",
+            ["@DB.SCHEMA.REPO/branches/main/s1.sql"],
+        ),
+        (
+            "@DB.schema.REPO/branches/main/a/S3.sql",
+            "@DB.schema.REPO/branches/main/",
+            ["@DB.schema.REPO/branches/main/a/S3.sql"],
         ),
     ],
 )
@@ -469,9 +497,68 @@ def test_execute(
 ):
     mock_execute.return_value = mock_cursor(
         [
-            {"name": "repo/branches/main/a/s3.sql"},
+            {"name": "repo/branches/main/a/S3.sql"},
             {"name": "repo/branches/main/s1.sql"},
             {"name": "repo/branches/main/s2"},
+        ],
+        [],
+    )
+
+    result = runner.invoke(["git", "execute", repository_path])
+
+    assert result.exit_code == 0, result.output
+    ls_call, *execute_calls = mock_execute.mock_calls
+    assert ls_call == mock.call(f"ls {expected_stage}", cursor_class=DictCursor)
+    assert execute_calls == [
+        mock.call(f"execute immediate from {p}") for p in expected_files
+    ]
+    assert result.output == os_agnostic_snapshot
+
+
+@pytest.mark.parametrize(
+    "repository_path, expected_stage, expected_files",
+    [
+        (
+            "@repo/branches/main/",
+            "@repo/branches/main/",
+            [
+                "@repo/branches/main/S2.sql",
+                "@repo/branches/main/s1.sql",
+                "@repo/branches/main/a/s3.sql",
+            ],
+        ),
+        (
+            "@repo/branches/main/s1.sql",
+            "@repo/branches/main/",
+            ["@repo/branches/main/s1.sql"],
+        ),
+        (
+            "@repo/branches/main/S2.sql",
+            "@repo/branches/main/",
+            ["@repo/branches/main/S2.sql"],
+        ),
+        (
+            "@repo/branches/main/a/s3.sql",
+            "@repo/branches/main/",
+            ["@repo/branches/main/a/s3.sql"],
+        ),
+    ],
+)
+@mock.patch(f"{STAGE_MANAGER}._execute_query")
+def test_execute_new_git_repository_list_files(
+    mock_execute,
+    mock_cursor,
+    runner,
+    repository_path,
+    expected_stage,
+    expected_files,
+    os_agnostic_snapshot,
+):
+    mock_execute.return_value = mock_cursor(
+        [
+            {"name": "/branches/main/s1.sql"},
+            {"name": "/branches/main/S2.sql"},
+            {"name": "/branches/main/a/s3.sql"},
         ],
         [],
     )