+++ RELEASE-NOTES.md
@@ -25,7 +25,8 @@
 
 ## Fixes and improvements
 * Duplicated keys in `snowflake.yml` are now detected and reported.
-
+* Fixed git execute not working with upper case in directory name.
+* Fixed `snow git setup` command behaviour for fully qualified repository names.
 
 # v3.0.0
 ## Backward incompatibility
@@ -59,7 +60,6 @@
 * Fix error handling and improve messaging when no artifacts provided.
 * Improved error message for incompatible parameters.
 * Fixed SQL error when running `snow app version create` and `snow app version drop` with a version name that isn't a valid Snowflake unquoted identifier
-
 
 # v2.8.0
 ## Backward incompatibility